requirements.txt
botbuilder-integration-aiohttp>=4.14.0
langchain
tabulate
openai
pandas
pyodbc

